+++ src/editor_extensions/conceal.ts
