+++ src/editor_extensions/conceal_fns.ts
