+++ src/editor_extensions/highlight_brackets.ts
